requirements.txt
pandas==1.5.3
py-slippi==1.6.2